move machine into MachineShopSimulator as an inner class @Xaivteev
